remove unused PluginRegistrySystemProps class
for GP-6213, removing System.getProperty calls
